test: test

+++ app/src/main/java/com/example/petsapce_week1/vo/HomeResponse.kt
@@ -10,7 +10,7 @@ data class HomeResponse(
 
 data class Result(
     val availableDays: List<String?>?,
-    val averageReviewScore: Int?,
+    val averageReviewScore: Float?,
     val city: String?,
     val district: String?,
     val numberOfReview: Int?,
